move InstUpdater to its own file
